Backend/requirements.txt
cffi==1.14.5
click==8.1.3
colorama==0.4.5
cryptography==3.4.7
Flask==2.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-serial==1.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.3
pycparser==2.20
PyMySQL==1.0.2
pyserial==3.5
python-dotenv==0.17.1
six==1.15.0
SQLAlchemy==1.4.11
typing==3.7.4.3
Werkzeug==2.1.2
zipp==3.8.0
